Header and footer .gsp working for loading and login
Note addition of aurelia-app=“main” in header-and-footer was necessary
to get it to work for the loading page.
We are still left with two different header and footer html, one in
header-and-footer.gsp and one set in header.html/js and footer.html/js
on the aurelia side. This is OK because login and loading have
different functionality in the header/footer.