Concatenate the image card code for the 5 images in a row in index.html
I wanted to get the html for the card display into one section for each
image so I created another nested repeat.for loop

to index into row.item1 using row.item[1] instead

so I had to change the way SearchableItemcontroller.groovy created the
json

I added a field called numberImagesThisRow so that we do not have the
empty missing cards at the bottom of the grid

Seems to work well on the proper designers, but not on the paco and
ralph ones. Not sure why, but did not bother to debug.