Trying to get select2 and editSearchableitem to work
Still buggy for location and not saving values.
Select2 items are NOT saving to JSON which means that they are not
attaching to the selectedSample type .
Also location is causing some extra activity in selection where it
selects again but always “select” or the lowest value.
Can fix this by selecting on name not id. But then need to change the
html/js side