Converted Type, Material and New Color fields to use select-control
I also renamed the "select2" custom element to "select-control" to avoid
some weird styling collision with the element "select2". I wasn't
entirely sure what particular property each of the three fields should
set on the model, so, that part may need to be tweaked in the callback
method for each of the 3 select boxes.