Created a select-address component and made it a global component that can be resued
Implemented an addressDatastore and a userDatastore, along with a ds
(datastore) manager.

Updated codebase to no longer call into the userService to getUser().
Now the userDatastore will call the userService.getUser(), and the
userService will update the datastore anytime the user object changes.

Added the select-address control to the editSampleRequest page.  Appears
to work client side, i.e. we send the updated address to the server, yet
the server responds sucessful but returns the old address (i.e. it's not
saving the new address)

More work may need to be done on the select-address control.  Currently
it attempts to get your addresses, provides you with a display of the
selected address, and an add button to add a new address.  Since this
was copied from the brand use case, we may need to add more options to
make it truly generic / global.