Setting smaller size for minimum height for the grey overall
Can’t make it smaller than this, as text will not fit. Fixes the
grid-shake (for all images I’ve checked and the ones that caused the
issue in the first place). So a compromise to fix grid shake, but some
overlap into white space when the image is shorter than 150px (which is
rare).